Put data-transfer objects (or query-response models) alongside the queries themselves
