--- backend/src/Models/Database.cs
@@ -1,15 +1,0 @@
-using System;
-
-namespace Database.Models;
-
-public sealed record Database(
-     Guid Uuid,
-     string Name,
-     string Description,
-     Uri Locator,
-     DatabaseVerificationState VerificationState,
-     string VerificationCode,
-     DatabaseOperatorEdge Operator,
-     bool CanCurrentUserUpdateNode,
-     bool CanCurrentUserVerifyNode
-);
