--- backend/src/Models/DatabaseOperatorEdge.cs
@@ -1,5 +1,0 @@
-namespace Database.Models;
-
-public sealed record DatabaseOperatorEdge(
-    Institution Node
-);
