--- backend/src/Models/DatabaseVerificationState.cs
@@ -1,7 +1,0 @@
-namespace Database.Models;
-
-public enum DatabaseVerificationState
-{
-    PENDING,
-    VERIFIED
-}
