--- backend/src/Models/Institution.cs
@@ -1,7 +1,0 @@
-using System;
-
-namespace Database.Models;
-
-public sealed record Institution(
-     Guid Uuid
-);
